+++ src/components/cardList.js
@@ -1,0 +1,4 @@
+export const createCardList = () => `
+  <ul class="trip-days"></ul>
+`;
+
