+++ src/components/menu.js
@@ -1,0 +1,7 @@
+export const createMenu = () => `
+  <div class="trip-info__main">
+    <h1 class="trip-info__title">Amsterdam</h1>
+    <p class="trip-info__dates">Mar 18&nbsp;—&nbsp;21</p>
+  </div>
+`;
+
