+++ src/constants.js
@@ -1,0 +1,12 @@
+const TRIPS_NUMBER = 3;
+const tripInfo = document.querySelector(`.trip-main__trip-info`);
+const tripControls = document.querySelector(`.trip-main__trip-controls`);
+const tripEvents = document.querySelector(`.trip-events`);
+
+
+export {
+  TRIPS_NUMBER,
+  tripInfo,
+  tripControls,
+  tripEvents
+};
