+++ src/utils.js
@@ -1,0 +1,8 @@
+const render = (container, template, place) => {
+  container.insertAdjacentHTML(place, template);
+};
+
+
+export {
+  render
+};
